% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather.R
\docType{data}
\name{feather_palettes}
\alias{feather_palettes}
\title{List of colour palettes}
\format{
An object of class \code{list} of length 5.
}
\usage{
feather_palettes
}
\description{
List of colour palettes
}
\keyword{datasets}
