% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feathers.R
\docType{data}
\name{feathers_palettes}
\alias{feathers_palettes}
\title{List of colour palettes}
\format{
An object of class \code{list} of length 12.
}
\usage{
feathers_palettes
}
\description{
List of colour palettes
}
\keyword{datasets}
