% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather.R
\name{get_pal}
\alias{get_pal}
\title{Select feather palette}
\usage{
get_pal(palette_name)
}
\arguments{
\item{palette_name}{Name of desired palette}
}
\value{
A vector of hex colour codes
}
\description{
This function returns colour palettes based on Australian birds
}
\examples{
get_pal("spotted_pardalote")
get_pal("bee_eater")
}
